Extract lerp to helper
